ID3 tag encoding detection
ID3v1 only supported ISO 8859-1 (Latin-1) texts, so many files were
created with other text encodings (Big5, Shift-JIS, etc). This is
still done for ID3v2 files, even if UTF-[8,16] is supported. It is
therefore necessary to try to detect which encoding each tag is
using.

The confidence of ID3 tag encoding detection is according to
the current system locale.

Change-Id: I9ce6f64c257f6a7324f331e10a16b0a50ec229d7